+++ src-d/diva-dns/DivaServer.cs
@@ -101,43 +101,7 @@ namespace diva_dns
         }
     }
 
-    /// <summary>
-    /// Simple post request for a Transaction. Send and wait for response with timeout.
-    /// </summary>
-    public class PostRequest
-    {
-        private readonly string _url;
-        private readonly HttpClient _client;
-        private readonly Transaction _transaction;
-
-        public HttpResponseMessage? ResponseMessage { get; private set; }
-
-        public PostRequest(HttpClient client, string url, Transaction transaction)
-        {
-            _client = client;
-            _url = url;
-            _transaction = transaction;
-        }
-
-        public bool SendAndWaitForReponse(int timeout_in_ms = -1)
-        {
-            try
-            {
-                var task = _client.PostAsJsonAsync(_url + "transaction/", _transaction);
-                if (task.Wait(timeout_in_ms))
-                {
-                    ResponseMessage = task.Result;
-                    return true;
-                }
-            }
-            catch (Exception e)
-            {
-                // Todo exception handling
-            }
-
-            return false;
-        }
-    }
+    
 
     public class DivaServer
     {
